+++ Flow/src/utopia/flow/async/process/TimedTasks.scala
@@ -82,6 +82,28 @@ class TimedTasks(waitLock: AnyRef = new AnyRef, shutdownReaction: ShutdownReacti
 		if (state == Completed)
 			runAsync()
 	}
+	
+	/**
+	  * Adds a new timed task. Restarts this process if it had completed already.
+	  * Note: Doesn't start this process if this process has not yet started.
+	  * @param task A task to add to the performed tasks
+	  */
+	def +=(task: TimedTask): Unit = add(task.firstRunTime) { task.run() }
+	/**
+	  * Adds new timed tasks. Restarts this process if it had completed already.
+	  * Note: Doesn't start this process if this process has not yet started.
+	  * @param tasks Tasks to add to the performed tasks
+	  */
+	def ++=(tasks: IterableOnce[TimedTask]): Unit = {
+		val iter = tasks.iterator
+		if (iter.hasNext) {
+			iter.foreach { t => _addTask(t.firstRunTime, () => t.run()) }
+			// Restarts if necessary (but not if broken)
+			if (state == Completed)
+				runAsync()
+		}
+	}
+	
 	/**
 	  * Adds a new task. Restarts this process if it had completed. The specified task is only run once.
 	  * @param time Time when the task should be run